+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/icon/DefaultIconService.java
@@ -104,7 +104,7 @@ public class DefaultIconService implements IconService {
 
   @Override
   @Transactional(readOnly = true)
-  public long getIconCount() {
+  public int getIconCount() {
     return customIconStore.getCustomIconCount();
   }
